jenkins: alert old patches jenkins job
This script queries gerrit for old patches that were remained untouched
for more than 30 days. (no updates, no patchsets etc')
and send the commiter a reminder email
this script will run in a jenkins job once a week.
if the patch is older than 90 days it will warn the commiter that the
patch will be abandoned automatically.

Change-Id: Ic50d7952f1efd904c4294aa5232cec716977ad66
Signed-off-by: Ohad Basan <obasan@redhat.com>