Added workspace cleanup job
Removes all the workspaces from all the slaves (even offline ones) that are not
currently used by any job.

It also cheks that there are no stalled mounts on the workspace and tries to
umount before deleting.

Added also a nice summary in case some of them failed to cleanup.

Change-Id: I97ad249a432129ca0f8b176d9fa8dbe6b1bfb5b5
Signed-off-by: David Caro <dcaroest@redhat.com>