Adding first pipeline
Using workflow, for now, only for lago. The pipeline is triggered only
on merge for now and ends deploying the generated artifacts to the
lago-stapshot repo if the checks and builds passed for all the
distros.

Requires ovirt custom jjb>=1.2.1.dev43-1 (available in the ci-tools
repo)

This will create one snapshot repo for each pipeline run, suffixed
with the build tag (jenkins-${JOB_NAME}-${BUILD_NUMBER}) of the
pipeline

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1277650
Change-Id: Ie5d510ef6388c4cf178c31971dc05941a4a19d0a
Signed-off-by: David Caro <dcaroest@redhat.com>