Add jenkins builders for imgbased
Trigger on changes, build for Fedora and EL (comments in imbased
indicate that there's a bug on EL7 presently, but it's being
installed on EL images, so that may be old)

The linters explicitly check for python3-pyflakes, etc, though.
And they're called from the RPM specfile.

Are these eventually going to run on EL7?

Change-Id: Icb6e015464108cb0715381f69a9ce9d410ee657d
Signed-off-by: Ryan Barry <rbarry@redhat.com>