Remove the job to build the libxml gem
A previous patch added a job to build the RPM for the libxml Ruby gem,
as that was needed for the ne Ruby SDK. But it has been decided to avoid
this dependency in the SDK, writing in C the bridge between Ruby and
libxml, as a Ruby extension instead. This patch removes the job.

Change-Id: Ifb1f7546d9f5c1cf11d9a4031499221d9e5e650c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>