Don't run Java SDK 3.6 build jobs on master merge
This patch changes the existing Java SDK build job so that it doesn't
build the SDK when changes are merged to the master branch. These builds
will be performed by other jobs introduced in the next patch.
Eventually, when there is no need to build the 3.6 Java SDK, this job
will be removed.

Change-Id: I27a81be8af242dd5dac441c4a7128171738cdbe3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>