Run Java SDK checks on EL7
Currently the checks for the the Java SDK only run in Fedora 23, because
the SDK is supposed to be platform independent. But in order to support
EL7 it needs to be patched, so that it works with the version of HTTP
components included in that distro. To ensure that there are no
regressions in that patching this patch changes the job so that it also
runs in EL7.

Change-Id: I6f21b636c29c362eedc23ef9c2f616c8c577eedd
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>