move artifactory-ovirt-org-settings.xml to jenkins repo
until now, this file was served staticly on the Jenkins master,
and each job that required it copied it. Few weeks ago it was added
to infra-puppet repositry. However, I think it better fits here, as this
is actually a job configuration and it will allow us better versioning
and visibility.

Each none standard-ci job that needs this file can clone it.

Change-Id: I917ec8d705f27f9525c564b6c38ac856a5711089
Signed-off-by: ngoldin <ngoldin@redhat.com>