Don't try to build Java SDK 4 using standard CI
There are currently two job configuration files for the Java SDK:

* ovirt-engine-sdk-java_create-rpms.yaml

  This configuration file defines the jobs that build version 3 of
  the SDK, and those jobs don't use the "automation" directory
  used by the oVirt "standard CI" mechanism. Instead of that the
  job just calls "make" to do the build.

* ovirt-engine-sdk-java_build-artifacts.yaml

  This configuration file defines the jobs that build version 4 of
  the SDK, using the "automation" directory.

But currently we are trying to build version 4 of the SDK (from branch
"sdk_4.0") from both job configuration files, and that will always fail
because that branch does't contain a Makefile that can be executed just
with the "make" command.

This patch changes the job configuration files of the SDK so that we
don't try to build version 4 of the Java SDK using the old approach
based on the "make" command.

Version 3 of the SDK will still be built the old way, as we don't expect
big changes there and we don't consider it worth the effort.

Change-Id: If4b8188ef5b68a1a94d5fdfa9375a7951cf49537
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>