Preventing unneeded check-patch jobs execution
When many patches are being pushed together, the
Jenkins queue becomes long and developers cannot
always waitfor the continous-integration +1 flag.
In that case, they are merging the patch before the
check-patch job runs.
In order to be efficient, I have added a short
bash script that checks if this execution is indeed
check-patch and not check-merged and if it is
it checks whether this patch had already been merged.
If it is, job with exit(0)

Change-Id: If9509e7280066327f0cca19a51555c4bee6a5dfa
Signed-off-by: Gil Shinar <gshinar@redhat.com>