Enable caching for packages in mock
Merge the package installation phase into the mock initializaiton phase
so that packages required by *.packages files will be cached along with
the rest of the chroot filesystem.

The drawback of this is that packages that are specified without full
versions will get updated only when the chroot cache is deleted. This
will also means that run on different slaves may end up with slightly
different package versions.

Change-Id: I6fb2cbdaf2da00ad979e546e66f18e5dabacce6e
Signed-off-by: Barak Korren <bkorren@redhat.com>