Added support for creating indexes of latest repos
The indexes are made to make it easy to figure out the URLs of the
latest snapshots for all repos.

Indexes are created in various formats to make them useful by client
software. Additionaly locks are used to ensure no two parallel attempts
to update the indexes are happening at the same time.

Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-575
Change-Id: I32aa9568595340a5a80b601c761e977ad3b0ba6f
Signed-off-by: Barak Korren <bkorren@redhat.com>