Fix standard-ci artifact collection
When a project's standard CI scripts run, they place artifacts in the
exported-artifacts directory under the project's git clone directory.
Jenkins, however, collects artifactes from the exported-artifacts
directory under $WORKSPACE.
To bridge this gap, there was some code in the standard_*.sh scripts
(That are in-charge of invoking mock_runner.sh itself) to move stuff
from the project's directory to the worspace directory.
That code only ran if mock_runner.sh (And hence the project's own CI
script) was successful.
This meant that artifact colection in case of failure was impossible.

This patch:
- Moves the artifact collection to a new script that is run as a
  post-build step.
- Removes the catch-all log collection code from cleanup.sh as that was
  probably an attempt to work-around the issue explained here.
- Made mock_runner.sh log to a path under exported-artifacts so its logs
  get collected
- Consolidates all standard-CI publishers into one standard-publisher
- Removed the unneeded extra publisher for the YAML DIFF jobs

Change-Id: Ife10ed15afb02333c9fbb92b362b376eafa0c047
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-988
Signed-off-by: Barak Korren <bkorren@redhat.com>