Made mock runner create logs in a safe place
Recently we made mock_runner.sh create logs in exported-artifacts to
make post-build collection scritps consistant. This can fail when the
script running inside mock tries to clean up exported-artifacts on it
own. In that case, mock_runner.sh can fail because the directories and
files it created are not longer there.

This patch moves log creation to a safe temporary directory, and moves
the logs to exported-artifacts only when mock_runner.sh is finished.

Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-992
Change-Id: I5f7f7756c3f44f4bdd6b5f6f40ee3b547b906fdb
Signed-off-by: Barak Korren <bkorren@redhat.com>