Changed ovirt version to 4.0 from 0.16
This change is needed because build artifacts jobs are
executing deploy to experimental job. The relevant deploy
to exporimental job name is being generated from a few parameters.
One of the in the version parameter.
No matter what's the package version, we'll always have
to "map" it into an oVirt version

Change-Id: I348542a93c2289183435f35d3547d258441279cd
Signed-off-by: Gil Shinar <gshinar@redhat.com>