Don't build RPMs for the JBoss modules Maven plugin
The Maven plugin that we use to build JBoss modules is a build-time
dependency that we use exclusively via Maven, and it doesn't need to be
packaged as an RPM package. This patch removes the jobs that build the
RPM packages, renames the .yaml file to use the '_standard' suffix
(insted of 'build-artifacts') and changes the remaining jobs so that
they only run in one distribution, as that is all what is needed to
verify that patches are acceptable.

Change-Id: I44bfd145bc4fe85d7b353011ee665f2ca0c2022f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>