experimental: append distro to repo path
Currently we download all distros, this is unneeded.

This patch 'hijacks' the 'distro' parameter, which it seems was intended
to be used for defining the 'chroot' distro. This should work because
we use el7 chroot, which happens to be what the VMs inside OST use.

However, if in the future we will decide to use fc* VMs inside OST,
we'd probably need to define another distro parameter, in order not to
restrict ourselves to running in fc* chroots too.

Fixes: https://ovirt-jira.atlassian.net/browse/OST-17

Change-Id: Id5e2e59da10b48c7b38598605fb0e6c7648f691b
Signed-off-by: Nadav Goldin <ngoldin@redhat.com>