image-ng: removed copyartifacts from job
The artifacts that are copied from the other jobs are not in use in this
script, because the script fetches the last successful builds for the
appliance and the node in prep_suite.

Also, when copyartifact tries to copy the appliance from master,
it searches for a job 'ovirt-appliance_ovirt-master_build-artifacts'
which doesn't exist, causing the image-ng job for master to fail.

This patch address these issued by removing copyartifacts from job.

Change-Id: Ia1f56e779a485299ed63feaafca5e3f7feaa9051
Signed-off-by: Yuval Turgeman <yturgema@redhat.com>