Make mirrors DNF (Fedora) compatible
DNF lookes for repo metadate files in the place pointed to by the
baseurl tag as opposed to next to the repomd.xml file which is where YUM
looks for them.

To make the mirrors compatible with DNF, the snapshots have been changed
to contain only the repomd.xml file, while all other metadata files are
stored in the mirror's 'base' directory.

To maintain YUM compatibility, HTTPD rewrite rules need to be configured
as specified in a comment in the mirror_mgr.sh script.

This patch also includes a change to the mirror_client.py script to make
it handle repos with 'metalink' configuration properly.

Change-Id: I9a20283ec74bb82b8d672389527edc99594f7012
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1112
Signed-off-by: Barak Korren <bkorren@redhat.com>