Make nightly publisher wait for cron job to finish
Sometimes we see a race-condition formed where the cron job that
processes the result of the nightly publisher job does not finish
running before the next publisher job for the same oVirt version runs.
This can happen if something in the publisher job output makes the cron
job fail constantly or if the publisher job is manually started
frequently enough.

This patch makes the publisher job wait for up to 50 minuts for the cron
job to finish and then fail under the assumption that we're seeing a
consistent cron job failure.

This will allow detecting the so-far transparent cron failures via the
Jenkins job failures

Also re-factored the publishers YAML to DRY it up

Change-Id: I875cc4aedc65580f07e517cc606a87a5f74355ea
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1161
Signed-off-by: Barak Korren <bkorren@redhat.com>