Set timestampts for packages we get from 'tested'
We found out an issue where RPMs we get from the 'OST tested' repo were
getting deleted by the repo cleanup script (clean_repos.sh). This is
because they were copied into the repo with the timestapms that match
when they were initially created (Or more accurately, uploaded into
experimental via the 'delpoy-to-experimental' job). This is different
from other RPMs that have the timestamp of when they were uploaded by
the publisher job.

This patch makes the publisher job timestamp the RPMs to the time in
which they were copied from the 'tested' repo.

Change-Id: I0cc5db81c2ec6062d4cf4486a7b9ca55a1f68587
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1167
Signed-off-by: Barak Korren <bkorren@redhat.com>