Final removal of master nightly snapshot copy jobs mech
This is the last patch in the series of removing all
package collection from master jobs using the nighly publisher.

After this patch is merged, the only way for packages to be
deployed to resources.ovirt.org is via the experimental flow.

This job however will continue to sync nightly snapshots
from the tested repo, making ovirt-snapshot a nightly
verified repo, as opposed to before of publishing non
verified arbritrary built packages.

Change-Id: I9bb49d6367e8d1efe8572736a68e5354f76aa8b0
Signed-off-by: Eyal Edri <eedri@redhat.com>