Enhance queue api to also return failure cause
Enhanced the API of the ChangeQueue objects to also return the specific
change that caused the failure when failed changes are detected.
(Up until now we just assumed the causing change will always be the
first in the list of changes that are removed due to failure)
This will allow adding more sophisticated cause detection code to be
added to the change queue objects.

Change-Id: Iaf34e047a0a53d3b6c151319f810431ccac73a3a
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1077