Creating a universal 'deploy-to' job
Job will trigger on Gerrit merges, wait for all build-artifacts jobs
that were triggered on the same event to finish and only then submit the
built packages to experimental.

This is aimed at solving a number of issues with the current
experimental deployment scheme, as described in the linked Jira tickets.

This job will also lay the groundwork to submitting builds as changes
to the change queue.

This patch also removes triggering of the older 'deploy-to' jobs from
Gerrit-based STD-CI jobs. This job does not yet support to current
configuration of GitHub-based jobs, so the older 'deploy-to' jobs are
kept for that.

Change-Id: I456d08c14e51f3278fa205d90360513c6de55545
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1315
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1077
Signed-off-by: Barak Korren <bkorren@redhat.com>