Added support for queued builds to deployer
Queued  builds are handled differently in Jenkins then running or
finished builds. This is why the deploy-to-experimental job was not
finding them.

This patch added support for finding queued builds and waiting for them
to get started.

For more details on the issue see Jira ticket.

Change-Id: I274eb227901f023bc2df9a52eead9d9f9cdd5f45
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1365
Signed-off-by: Barak Korren <bkorren@redhat.com>