Implement archive artifacts flow for containers
As part of our efforts for implement std-ci flow for containers,
I've added functionallity which is similar to 'archive artifacts' for
containers.
Every container which is tagged with 'exported-artifacts', will be
re-tagged with the $BUILD_URL's sha1 for further tracking, and pushed to our
intermediate phase repository in docker.io.

Also refactored collect_artifacts.sh not to be parsed by JJB.
Because the only reason we've passed collect_artifacts through JJB is to
get the project name and since it's being injected by the standard
template, we no longer need to do it.

Jira: https://ovirt-jira.atlassian.net/browse/OVIRT-1390

Change-Id: I0e44f17a8c63326a13991e7655590a6caf0f6250
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>