Add retries to mirror_client on JSON errors
We've seen some reoccurring failures where JSON decoding for information
from the mirrors server fails. This is probably because we're seeing a
race condition with the data update in the server.

This path adds retries so that when this happens we will re-download the
information from the server and hopefully get a full file.

Change-Id: I2f53126d0e2bd9421edaf57f019db9d5e520adc6
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1086