cleanup_slave: Fixed the cleanup of docker images
We can't cleanup docker images by image ID as we might remove
'whitelisted' images.
For example, a job that tags the official centos image with custom:tag,
then we'll search for all the images which doesn't belong to the
whitelisted repositories, and remove them (referring them by image id) -
this will cause the whitelisted image to be removed.

The new cleanup, is referring images by it's repository:tag, and will
not remove whitelisted images even if the repository name was retagged
by another image during the run.

Change-Id: I3c780cd5b7bc329e5715a1c7d9340340ac073cac