Inject 'tested' repos as mirrors for oVirt
In order to make OST always use the 'tested' repos as the fallback repos
when being run by the change queue, we use the mirror injection process
to inject the URLs for the 'tested' repos is if they were mirrors for
the oVirt released repos.

Also had to make some changes to mirror_client and mock_runner to enable
mock_runner to read mirrors data from files so that the mirrors
information will be downloaded at most once during the testing job

Fixed some typos in this patch as well.

Change-Id: Ifb045cc0db41800d10d6ce565d6b1744f1e15ce0
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1414