Remove patchset number from GerritMergedChange.id
The 'id' field for CQ changes is mainly for setting dependencies between
changes. When setting dependencies between Gerrit changes, the patchset
number that will be merged for a required change cannot always be known
at the point in time in which the requiring change is created. Therefore
we can't have the patchset number as part of the change id.

Change-Id: I3f7c5e37c7dafb472d3892fa20b38266f668babe
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1448
Signed-off-by: Barak Korren <bkorren@redhat.com>