Fix branches of metamodel jobs
Currently the API metamodel jobs with versions 'master' and '1.1' are
both triggered when is submitted to the 'master' branch. This patch
fixes that, so that the job with version 'master' is triggered only when
a change is submitted to the 'master' branch, and the job with version
'1.1' is triggered only when a change is submitted to the
'metamodel_1.1' branch.

Change-Id: I47004570fa3770c669b8d082c35d72ef70c222b5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>