try to build container if related files were changed
Since we've first introduced the option to build
containers in STD CI, the flow has been upgraded,
and today there are more files which are involved
with the flow.
I've updated check-patch to test the build of
container under a change to any of the files
which are related with the container build
and publishing flow.

Change-Id: I4af1afe78a309d37fb6c844a45b6e92bd0d06e41
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>