Adding injection of mirrors to slaves
We basically just overriding (yum|dnf).conf with a matching file from
the 'jenkins' repo that contains the URL of mirrors. We also have the
functionality to roll back this change if the job fails because of it.

Only adding repo configuration for CentOS 7 in this patch.

Change-Id: If11a435d5857ba120060847a784ed9882c5a00cd
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1468