Add parameters to system tests
System tests jobs used system-tests as jobs name
and ovirt-system-tests as hard coded project name.
We want to have the ability to use different project
name in different sites and, maybe, in the future to
support other types of products' system tests

pre jobs are not needed anymore. Removing them.

Change-Id: I2af96244e5819e7b1ccd368f48e2180ca39f35c2
Signed-off-by: Gil Shinar <gshinar@redhat.com>