Make STD-CI pipeline waste less nodes
STD-CI pipeline needs a node to initially clone sources to, get pipeline
code from and do STD-CI YAML parsing on. So far the pipelien kept that
node while it was running the actual STD-CI tasks on other nods.  This
path changes things so the node is released when it is not needed.

Change-Id: I0a6da849c075c805489fe5591c37af02e180c4d0
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1522