Apply timestamps and colors to all pipeline output
Change the pipeline-loader.groovy to make it apply timestamps and ANSI
colors to output from all parts of the pipeline.

Previously it only applied these setting to parts running on the node
the loader allocates for lading pipeline code and doing initialization
or single-node work because of the way older versions of these plugins
worked.

Change-Id: I9142df4a393a8e78fbb69408a157e461dec3d579
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1404