Prevent post merge Jobs from voting in Gerrit
This patch fixes a bug we've had where post merge jobs
from Jenkins failed to report results back to Gerrit.
The reason was that post merge failed jobs tried to vote
Continuous-Integration=-1 where down-voting is not allowed
post merge in Gerrit.

Now the vote result will be held in environmental variable
injected to the job, and the values hard coded, and only
chosen builders (currently check-patch) will have the ability
to vote.
This patch also fixes an issue we've had where build-artifacts
on demand jobs used to vote for CI +1.

Also, I've moved the Continuous-Integration voting values
from being configured in Jenkins to the YAMLs.

Jira:
[1] https://ovirt-jira.atlassian.net/browse/OVIRT-1502
[2] https://ovirt-jira.atlassian.net/browse/OVIRT-1156

Change-Id: I484cc04353d3cdeaef5b12c8eaaa5ba62baf4384
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>