Remove find bugs jobs
Since we have the find-bugs plugin configured and properly working
(and verified on ovirt-engine_master) for a while now, it's time to
remove the old find-bugs jobs since we don't support the template anymore.
Currently, the job is preventing other jobs which being triggered on
the same event from voting for the CI flag in Gerrit.
Since the only project that uses this job is ovirt-engine-4.1,
maintainers of the project will need to backport the find-bugs script
from ovirt-engine_master_check-patch to ovirt-engine_4.1_check-patch.

Change-Id: I552c30411b8be91363fcd56fe7638159f24d58e5
Jira: https://ovirt-jira.atlassian.net/browse/OVIRT-1635
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>