Added nested yaml config parser
The nested config parser can parse nested config files
from the form:
---
option-1: value-1
option-2:
  - value-2_1
  - value-2_2
     ...
  - value-2_n
option-3:
  - value-3:
      nested-option-4:
        - value-4
  ...

and returns a generator object that eventually iterates
over a unique list of vectors that are generated from
the given config.

The motivation for this patch is to eventually replace
the current automation/ directory with a single yaml config
that will be able to provide all the needed information
to std ci to generate jobs and setup the runtime environment.

Jira: https://ovirt-jira.atlassian.net/browse/OVIRT-1629
Change-Id: If0c311b878b2496683a6f525c52324e1ccb4e167
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>