Add post-merge triggering from GitHub
To complete STDCI support for GitHub, this patch adds the ability to
trigger jobs on GitHub push events.

Because of the way the available Jenkins plugins work, there are a few
jobs that need to be configured to enable this:
- A 'standard-webhook-trigger' job that triggers on all WebHook
  notifications from GitHub and triggers per-project jobs according to
  received parameters
- A per-project '*_standard-on-push' job that runs the actual STDCI
  functionality when it is triggered by 'standard-webhook-trigger'.

The per-project jobs could be regarded as a sort of a security
mechanism. Because anyone could trigger the webhook job, we need a way
to make sure we only run jobs for projects we intend to support. We will
create jobs only for projects we want to support. For other projects the
webhook event will simply be ignore.

Change-Id: I675164df76f5eb2d3ec8b78c1372576d90c8ca21
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1599