Added submission of GitHub changes to CQ
With this patch its possible to specify map project branches to change
queues in the automation.yaml file so that when changes are pushed to
those branches they are submitted to the specified queues.

This patch essentially implements for GitHub/pipeline jobs what
standard-enqueue implements for Gerrit/Freestyle jobs.

This patch requires that the following pathc will also be merged in
order to work reliably:
- https://gerrit.ovirt.org/c/82168

Change-Id: Ic7317d667d30b27fc0e811f1140b287b667dbf45
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1486
Signed-off-by: Barak Korren <bkorren@redhat.com>