Generalized CQ tester pipeline code
The change-queue-tester job have a lot of oVirt specific code in it that
was mixed with other generic change-queue related code. This patch
splits things up so that the oVirt specific code is in its own file.

This allows us to more easily create change queues for things that are
not oVirt. As an example this patch also adds code for the 'dummy'
change queue that basically does nothing.

Change-Id: I98cc741233e3c3fe059ca027cef230ffbd5dbb3b
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1534
Signed-off-by: Barak Korren <bkorren@redhat.com>