Created pusher.py, an automated push tool
This is moving the remaining functionality from
jobs/confs/python-scripts/poll-upstream-sources.py that is not already
handled by usrc.py to its own file that is usable as a stand-alone tool.

This is generally to allow using this functionality from piplines.

Some of the code needed to be re-factored to allow writing tests for it,
other parts were left as they were, with bugs included...

Change-Id: I3fafb7fed6987c4a6bc5f75a50e398db1ac7ef59
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1746
Signed-off-by: Barak Korren <bkorren@redhat.com>