Use oVirt version for names of metamodel jobs
The jobs that check and build the API metamodel have the version of the
project in the jobs: 1.0, 1.1, etc. But the CI system expects to have
the version of the oVirt project instead: 4.0, 4.1, etc. This patch
adjust the names of the jobs so that it will match what the CI system
expects.

Change-Id: I98c1a5d98bf4b9abf15a13cf60485b08cdf4a42c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>