Make setup and cleanup scripts work without sudo
Make the scripts able to work pass successfully even if sudo is not
configured on the slave for the jenkins user.

Also make all sudo calls everywhere use the '-n' parameter so that it
will fail when lacking permissions rather then waiting forever for the
user to enter a password.

Also improved logging a bit for some of the longer scripts.

This was done in order to enable using loaned slaves that are managed by
other people like the s390x slave.

Change-Id: I58eb7698352d2e405a87de4b6b8f3661034313f5
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1772
Signed-off-by: Barak Korren <bkorren@redhat.com>