ioprocess: Reorganize by version/stage
Simplify the project file by having one project per version, eliminating
the unmaintainable excludes. With this change, it should be easy to have
builds for all distros and architectures, and hard to forget to build
for some combination.

The new configuration is organized by version. Once we created a project
for released version we never have to change it. Changes are needed only
int master project. When a version is not maintained any more, can
remove the entire project.

These are the behavior changes introduced by this change:

- Jobs for 3.6 (0.15) were removed, we don't test this version for long
  time.
- check-merged jobs were removed, ioprocess check-merge.sh is a symlink
  to check-patch.sh.
- master fc25 jobs removed, fc25 is EOL now
- master fc26 jobs for ppc64le added, I don't know why there were
  missing, I guess this is a result of the complicated excludes.

Change-Id: I33c8ac00817b2f5b38bf91846c0fbafb3cbc511e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>