Add automerge permission checking to pusher.py
So far we've just checked that the 'automerge' commit message header is
set. Now we also check that the patch owner is actually allowed to use
the automerge functionality.

For a person to be allowed to use automerge, hes username most either be
explicitly specified in the list of maintainers in the push map file, or
he must be a member of one of the maintainer groups listed in the push
map file.

To implement this functionality, the PushDetails object had been
re-factored to be a fully fledged clas rather then simply a NamedTuple.

Change-Id: Idfa31522c1c269b74b24e7467e1123d8eec87f8d
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1474
Signed-off-by: Barak Korren <bkorren@redhat.com>