Enable mock to use bootstrap chroot for FC27/RAW
Since Fedora started switching to DNF, We had issues with setting up
mock for Fedora on CentOS hosts, because fedora packages were meant to
be installed via DNF rather then YUM. In Fedora 27 and RAWHIDE the
ability to use YUM finally broke completely.

Recent versions of mock offers a solution for this in the form of a so
called 'bootstrap chroot'. The idea is to setup a minimal Fedora chroot
frmo which DFN can be invoked to setup the final chroot for use.

This patch attempts to enable the 'bootstrap chroot' for use when
setting up Fedora RAWHIDE environments.

For this patch to work, mock version needs to be 1.4.9-1 and above.

Change-Id: Id62bb1ddf9160d3b135847c2bb85b89bbe1703c3
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1830