Enable customizing the server we deploy tested to
So far resources.ovirt.org had be hard-coded as the server that tested
packages are deployed to. Now we can customize it so that the code can
be used in other systems.

Also renamed the 'deploy' job template so that jobs for deploying
non-ovirt projects could be made using it.

Change-Id: I8de2c07174831895d113c6bd042de20abfac4cfb
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1474
Signed-off-by: Barak Korren <bkorren@redhat.com>