Make change-queue-tester not wait for deployment
Currently the change-queue-tester job for oVirt waits for the
deploy-to-tested job to finish before terminating and letting the change
queue schedule the next test. This can cause quite a delay since
deploy-to-tested can take quite a long time to finish.

This patch speeds things up by making change-queue-tester launch the
deploy-to-tested job asynchronously and have the next
change-queue-tester run wait for it to finish.

Change-Id: I7eecd337478e207015ddea92323c207d9cef8a81
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1837