Introduce Gerrit pipelines
This patch adds the support for Gerrit projects to use our
standard-pipeline.

The pipeline approch for Gerrit projects is same as for GitHub:
Users have to add their project to a single yaml called
'standard-pipelines' under projects/ dir and no extra configuration is
needed.

Since some Gerrit projects require credentials support, I implemented
this feature in this patch as well.

This patch also configures jenkins_standard-pipelines in staging-jenkins
so we'll have CI for this patch as well.

Jira-Ticket:
  -  https://ovirt-jira.atlassian.net/browse/OVIRT-1680
  -  https://ovirt-jira.atlassian.net/browse/OVIRT-1672

Change-Id: Ie945a17efc7404a417ee56811696235fc06fb18d
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>