Limit projects that standard-enqueue triggers on
So far the standard-enqueue job was configured to trigger on all Gerrit
merge events. This makes sense in oVirt where it has its own Gerrit server.

When Gerrit servers are shared between multiple projects, with previous
configuration the standard-enqueue could trigger on projects that had
nothing to do with it.

To resolve this, the project filter for 'standard-enqueue' was made
configurable.

Change-Id: I3a3c420d7f31a6c7567ec537c2d3f5219935f93b
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1474
Signed-off-by: Barak Korren <bkorren@redhat.com>