ioprocess: Add project for 4.2
This is mostly copy of the master project, but I assume that the
master project will change in the future, while the 4.2 project will
not, so little duplication now will make future changes easier.

For 4.2 we support only el7, but I left the fc27 builds since ioprocess
should work on this distro.

Change-Id: I04766d9089a90caeeb9b79a91631c5a53c2a1034
Signed-off-by: Nir Soffer <nsoffer@redhat.com>