dsl: Fixed default value assertion bug
The assertion for script's default value was not right. It checked if
DefaultValue singleton is in the script's configuration file, but the
DefaultValue was under 'default_value' key so the assertion always
returned False.
This patch fixes the bug and the tests.

Also improved a bit the error message when configurations can not be
found.

Change-Id: I10449f098b646dbd18dc35e76d1d73f4a252ab08
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>