dsl api: Add get_threads_and_globals API
Added API to fetch list of threads and global config.
This is needed because the pipelines need to use the global options even
in case we dont have jobs to trigger for the stage.
Since the pipelines interact with the dsl through the
pipeline_dict formatter, I've had to modify it to match the new API.

Change-Id: I8c48cfa3ba0f67efba1be750496e82690fbb3421