std enqueue: Add env.GERRIT_TRIGGER_CI_VOTE_LABEL
The way we configured Gerrit Trigger is that we set 'manually' env var
that later will be expanded by the Gerrit Trigger and define the CI vote.
Since standard-enqueue may be triggered with other jobs, it's env may be
used to summarize the result for Gerrit Trigger. For this reason, we
must set GERRIT_TRIGGER_CI_VOTE_LABEL to prevent the expansion from
failing.

Change-Id: I1cbd976d28b42da120d22aed817ec51802983afc