ovirt-system-test: run network suite nightly
the network suite has already found several regressions. We do not want
to find them when we execute it mamually and sporadically. It needs to
happen continuously.

At some point in the future, if the suite is proven to be stable, and is
not too slow, I hope to add it to the test queue.

Change-Id: I37c6565bdb46822158f0ab5bf64ce7de46faa379
Signed-off-by: Dan Kenigsberg <danken@redhat.com>