Run `pusher.py` on non-Gerrit events
So far the code triggering `pusher.py` was skipped for job build that
were not triggered by Gerrit events. This made it not work at all for
the 'poll' jobs.

The reason `pusher.py` was not triggered is that the code making sure
its not pushing needlessly was dependant on parameters set by the Gerrit
plugin.

The issue is that `pusher.py` needs to know if the current HEAD is
different then the commit that was originally checked out by Jenkins.
The hard part is figuring out what the original commit was.

This patch addresses the issue by:
1. Making Jenkins tag the commit it check out
2. Passing the tag name to `pusher.py`
3. Making `pusher.py` able to parse annotated tags properly.

Change-Id: I22a0d1784851f5a714e69f41fb9350a9a6e0fc24