Add symlinks resolving capabilities to usrc
In order to provide 'real' change resolving capabilities to usrc.py,
it needs to know how to resolve symlinks in git repos.
With this patch, usrc.py can resolve symlinks between files in local
repo and in symlinks that are spread across downstream/upstream repos.
For example, downstream symlink that links to a file in upstream, and
the upstream file was changed.

Usage:
CLI:
    jenkis/scripts/usrc.py changed-files --resolve-symlinks
Python:
    get_modfiied_files(new_commit, old_commit, resolve_links=True)

This is done as a preperation for conditional execution in STDCI DSL
(run stage/substage if file changed)

Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1894
Change-Id: I0ab982f14d6e60202aef64cc19c70067f34e8851
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>