Run builds of API model in Fedora Rawhide
This pach changes the jobs that build the API model so that they run in
Fedora Rawhide only. The reason for that is that the build
needs the 'publican' tool, and the version in CentOS (3.2.0) doesn't
work correct with DockBook 5. Since no artifact is consumed at all
from these jobs, running on Rawhide should ensure we are using latest
publican always.

Change-Id: Id5a8ea36dd03ace361f2e00a1756e2412e23b232
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>