Introduce conditional execution in STDCI DSL
Allow uses of stdci to configure conditional execution in their
configuration. Threads that will not satisfy the condition will be
excluded from parse.
Currently only supported condition is 'filechanged' and utilizes usrc.py
and it checks if a file was included in the commit.

Change-Id: I83763629ced906b8d5d9af616affc42d9f9c9fb6
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1894
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>