Modify mergeoptions in parser utils to merge runif
To properly merge and combine runif conditions from different depths of the
config as well as conditions from different vectors during the parse,
we combine every 2 different conditions under one 'all' condition.

'all' is an operator representing 'AND'. Hence, it requires all
of the conditions it warps to evaluate to True in order to evaluate to
True.

'any' is an operator representing 'OR'. Hence, it requires any of the
conditions it wraps to evaluate to True is order to evaluate to True.

Change-Id: I4e0bcd186d1d48bf4c2a2ac91caf94bd22ec34f1
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1894
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>